backend/pyproject.toml
[project]
name = "vuurkorf-backend"
version = "1.0.0"
description = "AI-gedreven vuurkorf personalisatie backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "rembg>=2.0.56",
    "opencv-python-headless>=4.9.0.80",
    "Pillow>=10.3.0",
    "vtracer>=0.6.5",
    "lxml>=5.2.1",
    "ezdxf>=1.3.4",
    "svgpathtools>=1.6.1",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
